fixed the testing
